Generate new token with forgotten password API
