+++ src/auth/models/forgotten_password.interface.ts
@@ -1,0 +1,5 @@
+export interface ForgottenPassword {
+  email?: string;
+  newPasswordToken?: string;
+  timestamp?: Date;
+}
