--- src/auth/models/forgottenpassword.interface.ts
@@ -1,6 +1,0 @@
-
-export interface ForgottenPassword {
-    email?: string;
-    newPasswordToken?: string;
-    timestamp?: Date;
-  }
