+++ src/auth/services/student_auth.service.ts
@@ -1,4 +1,4 @@
-import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable, Logger } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { JwtService } from '@nestjs/jwt';
 import { from, Observable, of } from 'rxjs';
@@ -7,23 +7,24 @@ import { Repository, TransactionManager, UpdateResult } from 'typeorm';
 import { StudentUserEntity } from '../models/student_user.entity';
 import { StudentUser } from '../models/student_user.interface';
 import * as bcrypt from 'bcrypt';
-import { ForgottenPassword } from '../models/forgottenpassword.interface';
-import { ForgottenPasswordEntity } from '../models/forgottenpassword.entity';
+import { ForgottenPasswordEntity } from '../models/forgotten_password.entity';
 
 @Injectable()
 export class StudentAuthService {
-    hashPassword(password: string): Observable<string> {
-        return from(bcrypt.hash(password,12))
-    }
-    constructor(
-        @InjectRepository(StudentUserEntity)
-        private readonly studentRepository: Repository<StudentUserEntity>,
-        
-        @InjectRepository(ForgottenPasswordEntity)
-        private readonly forgottenPasswordRepository: Repository<ForgottenPasswordEntity>,
+  private readonly logger = new Logger(StudentAuthService.name);
 
-        private jwtService: JwtService,
-    ) {}
+  hashPassword(password: string): Observable<string> {
+    return from(bcrypt.hash(password, 12))
+  }
+  constructor(
+    @InjectRepository(StudentUserEntity)
+    private readonly studentRepository: Repository<StudentUserEntity>,
+
+    @InjectRepository(ForgottenPasswordEntity)
+    private readonly forgottenPasswordRepository: Repository<ForgottenPasswordEntity>,
+
+    private jwtService: JwtService,
+  ) { }
 
 
     doesUserExist(email: string): Observable<boolean> {
@@ -142,32 +143,30 @@ export class StudentAuthService {
          }),
        );
      }
-     createForgottenPasswordToken(forgottenPassword: ForgottenPassword): Observable<ForgottenPassword> {
-       const {email} = forgottenPassword;
-    
-       return this.doesUserExist(email).pipe(
-         tap((doesUserExist: boolean) => {
-           if (!doesUserExist)
-             throw new HttpException(
-               'This user does not exist.',
-                HttpStatus.BAD_REQUEST,
-             );
-         }),
-         switchMap(() => {
-           return from(
-             this.forgottenPasswordRepository.save({
-               email,
-               newPasswordToken: (Math.floor(Math.random() * (9000000)) + 1000000).toString(), //Generate 7 digits number,
-               timestamp: new Date(),
-               }),
-             ).pipe(
-               map((user: ForgottenPassword) => {
-                 return user;
-               }),
-             );
-          }),
-       );
-     }
+     createForgottenPasswordToken(email: string) {
+      return from(this.doesUserExist(email).pipe(
+        tap((doesUserExist: boolean) => {
+          if (!doesUserExist) {
+            this.logger.warn("Someone tried to generate forgotten password token for non-existant student user");
+            return;
+          }
+  
+          this.forgottenPasswordRepository.save({
+            email,
+            new_password_token: (Math.floor(Math.random() * (9000000)) + 1000000).toString(), //Generate 7 digits number,
+            timestamp: new Date(),
+          });
+  
+          this.logger.log("Succesfully generated forgotten password token");
+        }),
+        map(() => {
+          return {
+            "message": 'E-mail sent away with link to password',
+            "statusCode": '201'
+          };
+        })
+      ));
+    }
 
      updateStudentProfileById(id: number, student_id: number): Observable<UpdateResult> {
       const user: StudentUser = new StudentUserEntity();
@@ -176,4 +175,3 @@ export class StudentAuthService {
       return from(this.studentRepository.update(id, user));
     }
   }
-
